Add third test of stochastic bandits
This test checks to see if the agent correctly values exploration when
it is uncertain about the latent state. It should explore only when it
has enough time to reap the benefits of exploration, and after its first
exploration step, it should stay on the arm if it got a valuable prize,
and switch to the safe arm if it did not.