Add and write about more testing.
Further testing on the problems with the boundVOI agent. Full comments
are in the file. Notable results are that the agent, when recursing on
belief, underestimates the expected utility of arms by a large amount.
State recursion also produces an underestimation, but by a much smaller
amount. My suspicion is that the VOI bound is somehow making the agent
myopic.