Start schellingGame branch
This branch is for developing two-player games, and agents to play them.
At the moment, we focus on coordination games, although that will change
in the future. This first commit adds a simple example of such a game,
which should both illustrate the type of games that we focus on as well
as a rough model of the API for games. The difficult next thing will be
to make models of agents modelling agents modelling themselves.